extract repeated code into a custom hook
